fix(calendar): L3-5753 Added coverage for getTimeZone method
